requirements.txt
streamlit>=1.32
torch>=2.0
torchvision
opencv-python-headless
Pillow
